requirements_bk.txt
Django==3.1.6
django-heroku==0.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
gensim==3.8.3
gunicorn==20.0.4
newspaper3k
nltk
pandas
http://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
spacy
textblob
numpy
transformers
boto3
django-storages

